refactor: extracted comment input box to comment component
