+++ client/src/app/components/audit-report/audit-report.component.scss
@@ -126,33 +126,10 @@ mat-panel-description {
   cursor: pointer;
 }
 
-#sendicon{
-  color: map-get($alta-accent, 600);
-  font-size: x-large;
-}
-
-.comment-text {
-  background: map-get($alta-accent, 900);
-  border: #2aabd2;
-  border-radius: 5px;
-  padding: 1em;
-  margin: 1em;
-  word-wrap: normal;
-  color: white;
-  word-break: break-all;
-}
-
 .comment{
   height: inherit;
   min-width: 20em;
   max-width: 35em;
-}
-
-.comment-input {
-  width: 90%;
-  position: absolute;
-  bottom: 0px;
-  margin-left: 1em;
 }
 
 .comment-button {
