+++ client/src/app/components/audit-report/audit-report.component.ts
@@ -1,4 +1,4 @@
-import {Component, OnInit, ViewChild, AfterViewChecked} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {TableManagementComponent} from '../TableManagement.component';
 import {AuditReportService} from '../../services/audits/audit-report.service';
 import {ManageAuditsService} from 'src/app/services/audits/manage-audits.service';
@@ -16,7 +16,7 @@ import {saveAs} from 'file-saver';
   templateUrl: './audit-report.component.html',
   styleUrls: ['./audit-report.component.scss']
 })
-export class AuditReportComponent extends TableManagementComponent implements OnInit, AfterViewChecked {
+export class AuditReportComponent extends TableManagementComponent implements OnInit{
 
   id: any;
   body: any;
@@ -77,12 +77,7 @@ export class AuditReportComponent extends TableManagementComponent implements On
       this.setAuditData();
 
       this.setResultsData();
-      this.setCommentData();
     });
-  }
-
-  ngAfterViewChecked(){
-    this.scrollToBottom();
   }
 
   // TODO: Fix appropriate backend calls
@@ -189,10 +184,6 @@ export class AuditReportComponent extends TableManagementComponent implements On
         this.updatePage();
         this.dataSource.paginator = this.paginator;
         this.dataSource.sort = this.sort;
-        if (data.comments){
-          this.comments = data.comments;
-        }
-
       }
     );
   }
@@ -231,30 +222,7 @@ export class AuditReportComponent extends TableManagementComponent implements On
     );
   }
 
-  setCommentData(): void {
-    this.comments = [];
-    this.auditReportService.getComments(this.userService.getOrgId(), this.id).subscribe(
-      (data: any) => {
-        for (let i = 0; i < data.length; i++){
-          let newComment = {
-            author: data[i].author,
-            content: data[i].content,
-            timestamp: new Date(Date.parse(data[i].created_timestamp))
-          }
-          this.comments.push(newComment);
-        }
-      }
-    );
-  }
 
-  scrollToBottom = () => {
-    try {
-      let ch = document.getElementById("comment-history");
-      if (ch != null){
-        ch.scrollTop = ch.scrollHeight;
-      }
-    } catch (err) {}
-  }
 
   updatePage(): void {
     this.updatePaginator();
